% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBAcolors.R
\docType{data}
\name{exhibits}
\alias{exhibits}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 14.
}
\usage{
exhibits
}
\description{
Use \code{names(exhibits)} to view list of palette names.
Currently:
}
\keyword{datasets}
