% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBAcolors.R
\name{mba_colors}
\alias{mba_colors}
\title{Custom Palette Generator for MBA-inspired themes}
\usage{
mba_colors(exhibit, n_colors, type = c("discrete", "continuous"), rev = FALSE)
}
\arguments{
\item{type}{The type of palette to generate, either "discrete" or "continuous."}

\item{exhibits}{The name of the color palette to draw from.}

\item{num_colors}{The number of colors to include in the palette.}
}
\value{
A vector of colors for data visualization.
}
\description{
Custom Palette Generator for MBA-inspired themes
}
\examples{
mba_colors("mba", num_colors = 100, palette_type = "continuous")
mba_colors("Nautilus", num_colors = 5, palette_type = "discrete")

}
